feat: add goals and games module and delete other modules
